Refacored the SignalsDeleteService into 2 celery tasks. One for determining the signals that are in a state for more than X days and one for the deletion of a signal.
