--- app/signals/apps/services/tests/scenarios/signals/__init__.py
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2023 Gemeente Amsterdam
