--- app/signals/apps/signals/tests/management/commands/delete_signals/__init__.py
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2022 Gemeente Amsterdam
