+++ app/signals/apps/signals/tests/scenarios/features/delete_signals.feature
